Clean up component
